Unify telemetry driver for horus and nv14 targets.
