+++ radio/src/targets/horus/CMakeLists.txt
@@ -302,6 +302,7 @@ set(FIRMWARE_SRC
   targets/common/arm/stm32/extmodule_driver.cpp
   targets/common/arm/stm32/stm32_pulse_driver.cpp
   targets/common/arm/stm32/extmodule_serial_driver.cpp
+  targets/common/arm/stm32/telemetry_driver.cpp
   targets/common/arm/stm32/stm32_usart_driver.cpp
   targets/common/arm/stm32/rotary_encoder_driver.cpp
   targets/common/arm/stm32/pwr_driver.cpp
