+++ radio/src/targets/nv14/hal.h
@@ -209,9 +209,9 @@
 // Telemetry
 #define TELEMETRY_RCC_AHB1Periph        (RCC_AHB1Periph_GPIOD | RCC_AHB1Periph_GPIOJ | RCC_AHB1Periph_DMA1)
 #define TELEMETRY_RCC_APB1Periph        RCC_APB1Periph_USART2
-#define TELEMETRY_REV_GPIO              GPIOJ
-#define TELEMETRY_RX_REV_GPIO_PIN       GPIO_Pin_8  // PJ.08
-#define TELEMETRY_TX_REV_GPIO_PIN       GPIO_Pin_7  // PJ.07
+#define TELEMETRY_INV_GPIO              GPIOJ
+#define TELEMETRY_RX_INV_GPIO_PIN       GPIO_Pin_8  // PJ.08
+#define TELEMETRY_TX_INV_GPIO_PIN       GPIO_Pin_7  // PJ.07
 #define TELEMETRY_DIR_GPIO              GPIOJ
 #define TELEMETRY_DIR_GPIO_PIN          GPIO_Pin_13 // PJ.13
 #define TELEMETRY_GPIO                  GPIOD
@@ -221,6 +221,9 @@
 #define TELEMETRY_GPIO_PinSource_RX     GPIO_PinSource6
 #define TELEMETRY_GPIO_AF               GPIO_AF_USART2
 #define TELEMETRY_USART                 USART2
+
+#define TELEMETRY_DMA_Stream_RX         DMA1_Stream5
+#define TELEMETRY_DMA_Channel_RX        DMA_Channel_4
 #define TELEMETRY_DMA_Stream_TX         DMA1_Stream6
 #define TELEMETRY_DMA_Channel_TX        DMA_Channel_4
 #define TELEMETRY_DMA_TX_Stream_IRQ     DMA1_Stream6_IRQn
@@ -231,10 +234,11 @@
 
 #define TELEMETRY_DIR_OUTPUT()          TELEMETRY_DIR_GPIO->BSRRH = TELEMETRY_DIR_GPIO_PIN
 #define TELEMETRY_DIR_INPUT()           TELEMETRY_DIR_GPIO->BSRRL = TELEMETRY_DIR_GPIO_PIN
-#define TELEMETRY_TX_POL_NORM()         TELEMETRY_REV_GPIO->BSRRH = TELEMETRY_TX_REV_GPIO_PIN
-#define TELEMETRY_TX_POL_INV()          TELEMETRY_REV_GPIO->BSRRL = TELEMETRY_TX_REV_GPIO_PIN
-#define TELEMETRY_RX_POL_NORM()         TELEMETRY_REV_GPIO->BSRRH = TELEMETRY_RX_REV_GPIO_PIN
-#define TELEMETRY_RX_POL_INV()          TELEMETRY_REV_GPIO->BSRRL = TELEMETRY_RX_REV_GPIO_PIN
+#define TELEMETRY_TX_NORM()             TELEMETRY_INV_GPIO->BSRRH = TELEMETRY_TX_INV_GPIO_PIN
+#define TELEMETRY_TX_INV()              TELEMETRY_INV_GPIO->BSRRL = TELEMETRY_TX_INV_GPIO_PIN
+#define TELEMETRY_RX_NORM()             TELEMETRY_INV_GPIO->BSRRH = TELEMETRY_RX_INV_GPIO_PIN
+#define TELEMETRY_RX_INV()              TELEMETRY_INV_GPIO->BSRRL = TELEMETRY_RX_INV_GPIO_PIN
+
 // USB
 #define USB_RCC_AHB1Periph_GPIO         RCC_AHB1Periph_GPIOA
 #define USB_GPIO                        GPIOA
